requirements.txt
fastapi
uvicorn
jinja2
alipay-sdk-python==3.7.409